+++ docs/CHANGELOG.md
@@ -1,3 +1,4 @@
+- #1120 Improve error handling in SequencerRegistry.
 - #1108 improve error handling in `ProverIncentives` module.
 - #1104 Add `deposit` method to prover incentives module.
 - #1003 Removes the `BlobData` enum, requiring sequencers to post the `Batch` struct directly on chain.